+++ components/BlogSection.vue
@@ -7,7 +7,7 @@
       <div class="relative max-w-7xl mx-auto">
         <div class="text-center">
           <h2 class="text-3xl text-copy-primary leading-9 tracking-tight font-extrabold text-white-900 sm:text-4xl sm:leading-10">
-            From my blog
+            Recent articles from my blog
           </h2>
           <p class="mt-3 max-w-2xl mx-auto text-xl leading-7 text-gray-500 sm:mt-4">
             This is a place for me to write and explore ideas, tools, process, etc.
