+++ components/molecules/GitHubIcon.vue
@@ -1,24 +1,24 @@
 <template>
   <div>
-    <icon
-      href="https://github.com/iancleary"
-      title="Link to my GitHub Profile"
+    <external-link
+      link="https://github.com/iancleary"
+      alt="Link to my GitHub Profile"
     >
       <!-- https://tailwindcss.com/docs/fill#usage -->
       <!-- eslint-disable-next-line -->
-      <git-hub-icon size="1.5x" class="fill-current text-white h-8 w-8"></git-hub-icon>
-    </icon>
+      <git-hub-icon size="1.5x" class="fill-current text-white h-8 w-8" />
+    </external-link>
   </div>
 </template>
 
 <script>
 import { GitHubIcon } from 'vue-simple-icons';
-import Icon from '@/components/atoms/Icon.vue';
+import ExternalLink from '../atoms/ExternalLink.vue';
 
 export default {
   components: {
     GitHubIcon,
-    Icon,
+    ExternalLink,
   },
 };
 </script>
