+++ components/molecules/LinkedInIcon.vue
@@ -1,24 +1,24 @@
 <template>
   <div>
-    <icon
-      href="https://linkedin.com/in/ian-c-6566b22a"
-      title="Link to my LinkedIn Profile"
+    <external-link
+      link="https://linkedin.com/in/ian-c-6566b22a"
+      alt="Link to my LinkedIn Profile"
     >
       <!-- https://tailwindcss.com/docs/fill#usage -->
       <!-- eslint-disable-next-line -->
       <linked-in-icon class="fill-current text-white h-8 w-8"></linked-in-icon>
-    </icon>
+    </external-link>
   </div>
 </template>
 
 <script>
 import { LinkedInIcon } from 'vue-simple-icons';
-import Icon from '@/components/atoms/Icon.vue';
+import ExternalLink from '../atoms/ExternalLink.vue';
 
 export default {
   components: {
     LinkedInIcon,
-    Icon,
+    ExternalLink,
   },
 };
 </script>
