+++ components/molecules/TwitterIcon.vue
@@ -1,24 +1,24 @@
 <template>
   <div>
-    <icon
-      href="https://twitter.com/icancclearynow"
-      title="Link to my Twitter Profile"
+    <external-link
+      link="https://twitter.com/icancclearynow"
+      alt="Link to my Twitter Profile"
     >
       <!-- https://tailwindcss.com/docs/fill#usage -->
       <!-- eslint-disable-next-line -->
       <twitter-icon class="fill-current text-white h-8 w-8"></twitter-icon>
-    </icon>
+    </external-link>
   </div>
 </template>
 
 <script>
 import { TwitterIcon } from 'vue-simple-icons';
-import Icon from '@/components/atoms/Icon.vue';
+import ExternalLink from '../atoms/ExternalLink.vue';
 
 export default {
   components: {
     TwitterIcon,
-    Icon,
+    ExternalLink,
   },
 };
 </script>
